renamed postCardContainer component to postCard
