+++ src/components/postCard/index.js
@@ -1,4 +1,4 @@
-import PostCard from './container/postCardContainer';
+import PostCard from './container/postCard';
 
 export { PostCard };
 export default PostCard;
